Added 128 processes to the prime testing pool.
Kind of interesting how this scales. At 10 processes we are far from
saturating the CPU, at 32 processes we hit 50% utilization on all
threads, but can't increase that by raising the process count
higher. It is maxed out at 50%. Intriguingly all the erlang schedulers
are nearly maxed out, and there is one "dirty cpu" line in observer
that rises while the benchmark is working. Here is the benchmark:

lists:foreach(
  fun(N) ->
    spawn(fun() ->
      io:format("~p: ~p primes~n",
                [N, length(lists:filter(
                   fun(X) -> isaprime:test(X) end,
                   [rand:uniform(
                        99999999999999999999999999999999999999999999999)
                    || _ <- lists:seq(1, 500)]))])
     end)
   end,
   lists:seq(1, 10000)).

I could probably learn something by attaching a profiler or a trace to
this, is it scheduler overhead? is it the very large work queue? Is it
just that the work done by each tester is too small relative to the
time to send the response back through two processes? Is it the time while
each test server waits after it requests new work?

It will be interesting to see how the load balancer version does in
comparison, maybe it will be able to push higher utilization with many
smaller queues.

Lots to learn