First cut of load balancer and each teser keeping its own queue
Same performance as the previous version.

Since we need to be able to restart any given tester, this design
won't quite work. We need to keep the contents of the queue outside
the tester process. The best place to keep it is probably in the load
balancer, or an ets table. This gives the load balancer all the
information it needs to restart the servers and to keep track of the
load on each server.