Store the pending requests in an ets table
This will facilitate restarting workers when they crash by keeping the
pending requests alive in the load balancer so they can be
redistributed to other workers.

The performance seems to be pretty good, and I implemented counters
for each worker so the load can be known immediately. Given the
generally small size of the table, I'm not sure if this is worth
it; however, maybe with the frequent deletions from the table it will
perform better than using ets:select_count/2.

There is a lack of atomicity when the worker finishes a job since the
delete and the decrement of the counter are separate operations. This
leads to potentially sub-optimal load balancing decisions, but the
restart/work redistribution in the event of a failure can reset the
counter, ensuring that it does not become inconsistent.