Duplicate data across two nodes
Use mnesia to stor the data. All operations are dirty, since the usage
patterns mean violating isolation won't cause problems (at least on a
single node). The restart mechanism, however, needs more work. When
the load balancer server dies (or the entire node goes down) the
processes waiting for a response in gen_server:call/2 are killed. To
work around this and allow work to seemlessly be shifted to another
node with the only consequence of a node failing being potential
timeouts will require changing from a call to a cast and handling the
reply differently.

Not really what the exercise called for, but it should be fun and it
will give me the opportunity to try out the new(ish) alias/ bifs.