Rewrite isaprime:test/1 so clients tolerate load balancer failure
Change from a gen_server:call to a gen_server:cast so the caller does
not raise an exception if the server fails. Since the responses come
from the tester servers, any queued work can still be finished. This
may even work for a full node failure, but it seems unlikely that the
backup node will be up and running in time to prevent a lot of
timeouts.