Set up the application to run on two nodes
My plan with this was that the backup node could take over the work
from the primary server with the only consequence being a few
timeouts; however, the application isn't really designed to support
that. Since the load balancer server is registered locally, and the
API forces us to make all requests from the local node there isn't
really a point to trying to reply to processes that die with the
node. Furthermore, my restart mechanism uses is_process_alive/1 to
check the status of the tester servers, but that process only works on
local PIDs.

So, I'll abandon this approach for the node failure and for the
restart I'll just clear the mnesia tables and start with a clean
slate. This is kind of weird because there isn't much point to
replicating the data between two nodes in this case, but for a
different task (that doesn't require direct replies to the requesting
process).